# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: api
  version: 1.0.0
  description: EngineeringOS - Autonomous AI Engineering Platform API

# PR-02: Global security default — all /api/* routes require a valid Clerk
# session unless an operation explicitly overrides with `security: []`.
# This ensures new routes are protected by default without having to remember
# to add a per-operation security declaration.
security:
  - ClerkSession: []
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks with optional filters
      tags: [Tasks]
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TaskPriority"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      summary: Create a new task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/tasks/{taskId}:
    get:
      operationId: getTask
      summary: Get task by ID
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      summary: Update task status or fields
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      summary: Delete a task
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/tasks/{taskId}/execute:
    post:
      operationId: executeTask
      summary: Execute a task through the AI pipeline
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Task execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/tasks/{taskId}/retry:
    post:
      operationId: retryTask
      summary: Retry a failed task
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Task queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/tasks/{taskId}/rollback:
    post:
      operationId: rollbackTask
      summary: Roll back a task's changes
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/tasks/{taskId}/logs:
    get:
      operationId: getTaskLogs
      summary: Get execution logs for a task
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskLog"

  # ─── Rules ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: >
        Clerk session cookie. All /api/* routes (except /api/healthz) require a
        valid Clerk session. The cookie is set automatically by the Clerk frontend
        SDK after sign-in — no manual header injection is needed.
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, queued, running, verifying, completed, failed, cancelled]
    TaskPriority:
      type: string
      enum: [p0, p1, p2, p3]
...
    Task:
      type: object
      required: [id, projectId, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        ruleId:
          type: string
        workflowId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        relatedFiles:
          type: array
          items:
            type: string
        dependsOn:
          type: array
          items:
            type: string
        retryCount:
          type: integer
        maxRetries:
          type: integer
        phase:
          type: string
        prompt:
          type: string
        agentResponse:
          type: string
        verificationResult:
          $ref: "#/components/schemas/VerificationResult"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    CreateTaskInput:
      type: object
      required: [projectId, title, priority]
      additionalProperties: false
      properties:
        projectId:
          type: string
        ruleId:
          type: string
        workflowId:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        relatedFiles:
          type: array
          items:
            type: string
        dependsOn:
          type: array
          items:
            type: string
        phase:
          type: string
    UpdateTaskInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
    TaskLog:
      type: object
      required: [id, taskId, level, message, timestamp]
      properties:
        id:
          type: string
        taskId:
          type: string
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        metadata:
          type: object
        timestamp:
          type: string
          format: date-time
    VerificationResult:
      type: object
      required: [passed, steps]
      properties:
        passed:
          type: boolean
        steps:
          type: array
          items:
            type: object
            required: [name, passed]
            properties:
              name:
                type: string
              passed:
                type: boolean
              output:
                type: string

    # ─── Rule ──────────────────────────────────────────────────────────────────
...